Add init_sentry to acoustic_sync
Move init_sentry from app.py to new file exception_capture.py, and call
it during startup for acoustic_sync.py.
